fuzzing: move alpine utils to fuzz utils
